// web/src/proto/scribe/v1/item_connect.ts
// @generated by protoc-gen-connect-es v1.6.1 with parameter "target=ts"
// @generated from file scribe/v1/item.proto (package scribe.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { CreateItemRequest, CreateItemResponse, DeleteItemRequest, DeleteItemResponse, GetItemRequest, GetItemResponse, ListItemsRequest, ListItemsResponse, UploadItemImageRequest, UploadItemImageResponse } from "./item_pb.js";
import { MethodKind } from "@bufbuild/protobuf";

/**
 * @generated from service scribe.v1.ItemService
 */
export const ItemService = {
  typeName: "scribe.v1.ItemService",
  methods: {
    /**
     * @generated from rpc scribe.v1.ItemService.ListItems
     */
    listItems: {
      name: "ListItems",
      I: ListItemsRequest,
      O: ListItemsResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc scribe.v1.ItemService.GetItem
     */
    getItem: {
      name: "GetItem",
      I: GetItemRequest,
      O: GetItemResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc scribe.v1.ItemService.CreateItem
     */
    createItem: {
      name: "CreateItem",
      I: CreateItemRequest,
      O: CreateItemResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc scribe.v1.ItemService.UploadItemImage
     */
    uploadItemImage: {
      name: "UploadItemImage",
      I: UploadItemImageRequest,
      O: UploadItemImageResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc scribe.v1.ItemService.DeleteItem
     */
    deleteItem: {
      name: "DeleteItem",
      I: DeleteItemRequest,
      O: DeleteItemResponse,
      kind: MethodKind.Unary,
    },
  }
} as const;